// client/src/net/fbs/afps/protocol/game-event-type.ts
// automatically generated by the FlatBuffers compiler, do not modify

/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-explicit-any, @typescript-eslint/no-non-null-assertion */

export enum GameEventType {
  HitConfirmed = 0,
  ProjectileSpawn = 1,
  ProjectileRemove = 2
}